+++ packages/rnv-sdk-webpack/package.json
@@ -27,8 +27,8 @@
         "watch:all": "yarn watch & yarn test:watch"
     },
     "dependencies": {        
-        "better-opn": "3.0.1",
-        "axios": "0.24.0",
+        "better-opn": "1.0.0",
+        "axios": "0.21.2",
         "ip": "1.1.5",
         "command-exists": "1.2.9"
     },
